+++ nirum/rpc.py
@@ -1,29 +1,33 @@
 """:mod:`nirum.rpc`
 ~~~~~~~~~~~~~~~~~~~
 
+.. deprecated:: 0.6.0
+   This module and all it has provided are deprecated or obsolete.  The most
+   of them are now distributed as separated packages, or replaced by a newer
+   concept.
+
+   It will be completely obsolete at version 0.7.0.
+
 """
-import argparse
 import collections
 import json
 import typing
+import warnings
 
 from six import integer_types, string_types
 from six.moves import urllib
 from werkzeug.exceptions import HTTPException
 from werkzeug.http import HTTP_STATUS_CODES
 from werkzeug.routing import Map, Rule
-from werkzeug.serving import run_simple
 from werkzeug.wrappers import Request as WsgiRequest, Response as WsgiResponse
 
-from .constructs import NameDict
 from .deserialize import deserialize_meta
-from .exc import (InvalidNirumServiceMethodNameError,
-                  InvalidNirumServiceMethodTypeError,
-                  NirumProcedureArgumentRequiredError,
+from .exc import (NirumProcedureArgumentRequiredError,
                   NirumProcedureArgumentValueError,
                   UnexpectedNirumResponseError)
-from .func import import_string, url_endswith_slash
+from .func import url_endswith_slash
 from .serialize import serialize_meta
+from .service import Service as BaseService
 
 __all__ = 'Client', 'WsgiApp', 'Service', 'client_type', 'service_type'
 JSONType = typing.Mapping[
@@ -31,38 +35,34 @@ JSONType = typing.Mapping[
 ]
 
 
-class Service(object):
-    """Nirum RPC service."""
+class Service(BaseService):
+    """Abstract base of Nirum services.
 
-    __nirum_service_methods__ = {}
-    __nirum_method_names__ = NameDict([])
+    .. deprecated:: 0.6.0
+       Use :class:`nirum.service.Service` instead.
+       It will be completely obsolete at version 0.7.0.
 
-    @staticmethod
-    def __nirum_method_error_types__(k, d=None):
-        return d
+    """
 
     def __init__(self):
-        for method_name in self.__nirum_service_methods__:
-            try:
-                method = getattr(self, method_name)
-            except AttributeError:
-                raise InvalidNirumServiceMethodNameError(
-                    "{0}.{1} inexist.".format(
-                        typing._type_repr(self.__class__), method_name
-                    )
-                )
-            if not callable(method):
-                raise InvalidNirumServiceMethodTypeError(
-                    "{0}.{1} isn't callable".format(
-                        typing._type_repr(self.__class__), method_name
-                    )
-                )
+        warnings.warn(
+            'nirum.rpc.Service is deprecated; use nirum.service.Service '
+            'instead.  It will be completely obsolete at version 0.7.0.',
+            DeprecationWarning
+        )
+        super(Service, self).__init__()
 
 
 class WsgiApp:
     """Create WSGI application adapt Nirum service.
 
     :param service: A nirum service.
+
+    .. deprecated:: 0.6.0
+       Use ``nirum_wsgi.WsgiApp`` (provided by `nirum-wsgi
+       <https://github.com/spoqa/nirum-python-wsgi>`_ package) instead.
+
+       It will be completely obsolete at version 0.7.0.
 
     """
 
@@ -73,6 +73,12 @@ class WsgiApp:
     ])
 
     def __init__(self, service):
+        warnings.warn(
+            'nirum.rpc.WsgiApp is deprecated; use nirum_wsgi.WsgiApp '
+            '(provided by nirum-wsgi package).  It will be completely '
+            'obsolete at version 0.7.0.',
+            DeprecationWarning
+        )
         self.service = service
 
     def __call__(self, environ, start_response):
@@ -323,8 +329,23 @@ class WsgiApp:
 
 
 class Client(object):
+    """HTTP service client base class.
+
+    .. deprecated:: 0.6.0
+       Use :class:`nirum.transport.Transport` and
+       :mod:`nirum_http.HttpTransport` (provided by `nirum-http
+       <https://github.com/spoqa/nirum-python-http>` package) instead.
+       It will be completely obsolete at version 0.7.0.
+
+    """
 
     def __init__(self, url, opener=urllib.request.build_opener()):
+        warnings.warn(
+            'nirum.rpc.Client is deprecated; use nirum.transport.Transport '
+            'and nirum_http.HttpTransport (provided by nirum-http package) '
+            'instead.  It will be completely obsolete at version 0.7.0.',
+            DeprecationWarning
+        )
         self.url = url_endswith_slash(url)
         self.opener = opener
 
@@ -431,21 +452,3 @@ class Client(object):
 # with postfix named `_type`
 service_type = Service
 client_type = Client
-
-
-def main():
-    parser = argparse.ArgumentParser(description='Nirum service runner')
-    parser.add_argument('-H', '--host', help='the host to listen',
-                        default='0.0.0.0')
-    parser.add_argument('-p', '--port', help='the port number to listen',
-                        type=int, default=9322)
-    parser.add_argument('-d', '--debug', help='debug mode',
-                        action='store_true', default=False)
-    parser.add_argument('service_impl', help='service implementation name')
-    args = parser.parse_args()
-    service_impl = import_string(args.service_impl)
-    run_simple(
-        args.host, args.port, WsgiApp(service_impl),
-        use_reloader=args.debug, use_debugger=args.debug,
-        use_evalex=args.debug
-    )